+++ NEWS.md
@@ -9,7 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Changed
 
-* bumped dependency of ManifoldsBase.jl to 0.16, split `exp` into `exp` and `exp_fused` accordingly
+* bumped dependency of ManifoldsBase.jl to 1.0, split `exp` into `exp` and `exp_fused` accordingly
+* ManifoldsBase.jl 1.0 also moved from `TVector` to `Tangentvector`s in type names.
+  The following names are adapted
+    * Renamed `HyperboloidTVector` (now deprecated) to `HyperboloidTangentVector`
+    * Renamed `OrthogonalTVector` (now deprecated) to `OrthogonalTangentVector`
+    * Renamed `PoincareBallTVector` (now deprecated) to `PoincareBallTangentVector`
+    * Renamed `PoincareHalfSpaceTVector` (now deprecated) to `PoincareHalfSpaceTangentVector`
+    * Renamed `ProjectorTVector` (now deprecated) to `ProjectorTangentVector`
+    * Renamed `StiefelTVector` (now deprecated) to `StiefelTangentVector`
+    * Renamed `TuckerTVector` (now deprecated) to `TuckerTangentVector`
+    * Renamed `UMVTVector` (now deprecated) to `UMVTangentVector`
+* the internal access `array_value` is now `internal_value`, compare to its renaming in `ManifoldsBase`
+
 
 ## [0.10.12] - 2025-01-10